Se realizaron avances en la parte de excepciones
