+++ Algoritmos/README.md
@@ -1,1 +1,7 @@
-En este archivo se van a agregar los incisos de investigación y las instrucciones de ejecución de cada programa
+# IE-0217 Estructuras Abstractas de Datos y Algoritmos para Ingeniería
+# Tarea #1
+# Gabriel Barahona Otoya B70896
+
+
+## Instrucciones de ejecución de cada programa:
+### Analice cada algoritmo y determine la complejidad que estos tienen utilizando la notación Big O (O grande).
